+++ convert/src/main/kotlin/com/github/zeldigas/text2confl/convert/markdown/MarkdownConfiguration.kt
@@ -2,5 +2,6 @@ package com.github.zeldigas.text2confl.convert.markdown
 
 data class MarkdownConfiguration(
     val parseAnyMacro:Boolean = true,
-    val supportedMacros:List<String> = emptyList()
+    val supportedMacros:List<String> = emptyList(),
+    val emoji:Boolean = true
 )